move code from securityTagManager.js to patientScopeManager.js
