+++ src/operations/common/patientScopeManager.js
@@ -3,19 +3,26 @@ const { RethrownError } = require('../../utils/rethrownError');
 const { DatabaseQueryFactory } = require('../../dataLayer/databaseQueryFactory');
 const { PersonToPatientIdsExpander } = require('../../utils/personToPatientIdsExpander');
 const { ScopesManager } = require('../security/scopesManager');
+const Resource = require('../../fhir/classes/4_0_0/resources/resource');
+const { NestedPropertyReader } = require('../../utils/nestedPropertyReader');
+const { ForbiddenError } = require('../../utils/httpErrors');
+const { isUuid } = require('../../utils/uid.util');
+const { PatientFilterManager } = require('../../fhir/patientFilterManager');
 
 class PatientScopeManager {
-        /**
+    /**
      * constructor
      * @param {DatabaseQueryFactory} databaseQueryFactory
      * @param {PersonToPatientIdsExpander} personToPatientIdsExpander
      * @param {ScopesManager} scopesManager
+     * @param {PatientFilterManager} patientFilterManager
      */
     constructor (
         {
             databaseQueryFactory,
             personToPatientIdsExpander,
-            scopesManager
+            scopesManager,
+            patientFilterManager
         }
     ) {
         /**
@@ -35,6 +42,12 @@ class PatientScopeManager {
          */
         this.scopesManager = scopesManager;
         assertTypeEquals(scopesManager, ScopesManager);
+
+        /**
+         * @type {PatientFilterManager}
+         */
+        this.patientFilterManager = patientFilterManager;
+        assertTypeEquals(patientFilterManager, PatientFilterManager);
     }
 
     /**
@@ -65,7 +78,7 @@ class PatientScopeManager {
         }
     }
 
-        /**
+    /**
      * Gets Patient id from identifiers
      * @param {string} base_version
      * @param {string} personIdFromJwtToken
@@ -106,7 +119,7 @@ class PatientScopeManager {
      * @param {string[] | null} patientIdsFromJwtToken
      * @returns {Promise<string[]|null>}
      */
-    async getPatientIdsFromScope ({ base_version, isUser, personIdFromJwtToken, patientIdsFromJwtToken }) {
+    async getPatientIdsFromScopeAsync ({ base_version, isUser, personIdFromJwtToken, patientIdsFromJwtToken }) {
         /**
          * @type {string[]}
          */
@@ -124,6 +137,89 @@ class PatientScopeManager {
             : patientIdsLinkedToPersonId;
 
         return allPatientIdsFromJwtToken;
+    }
+
+    /**
+     * Gets value of patient property from resource
+     * @param {Resource} resource
+     * @return {string | undefined}
+     */
+    getValueOfPatientPropertyFromResource ({ resource }) {
+        assertTypeEquals(resource, Resource);
+        /** @type {string} */
+        const resourceType = resource.resourceType;
+        // Get name of patient property for this resource type
+        /**
+         * @type {string|string[]|null}
+         */
+        const patientFilterProperty = this.patientFilterManager.getPatientPropertyForResource({
+            resourceType
+        });
+        // If this resource has a patient property then get the value of that property
+        if (patientFilterProperty) {
+            if (Array.isArray(patientFilterProperty)) {
+                for (const p of patientFilterProperty) {
+                    // if patient itself then search by _uuid
+                    if (p === 'id') {
+                        return resource._uuid;
+                    } else {
+                        const propertyUuid = p.replace('.reference', '._uuid');
+                        const value = NestedPropertyReader.getNestedProperty({ obj: resource, path: propertyUuid });
+                        if (value !== undefined) {
+                            return value.replace('Patient/', '');
+                        }
+                    }
+                }
+            } else {
+                if (patientFilterProperty === 'id') {
+                    return resource._uuid;
+                } else {
+                    const propertyUuid = patientFilterProperty.replace('.reference', '._uuid');
+                    const value = NestedPropertyReader.getNestedProperty({ obj: resource, path: propertyUuid });
+                    if (value !== undefined) {
+                        return value.replace('Patient/', '');
+                    }
+                }
+            }
+        }
+        return undefined;
+    }
+
+    /**
+     * Gets whether the user can write the given resource with the given patient scope
+     * @param {string[] | null} patientIds
+     * @param {Resource} resource
+     * @return {boolean}
+     */
+    canWriteResourceWithPatientScope ({ patientIds, resource }) {
+        assertTypeEquals(resource, Resource);
+
+        /** @type {string} */
+        const resourceType = resource.resourceType;
+
+        // if this resource is not allowed to be written with a patient scope then return an error
+        if (!this.patientFilterManager.canAccessResourceWithPatientScope({ resourceType })) {
+            throw new ForbiddenError(`Resource type ${resourceType} cannot be written via a patient scope`);
+        }
+        // separate uuids from non-uuids
+        const patientUuids = patientIds.filter(id => isUuid(id));
+
+        /** @type {string} */
+        const patientForResource = this.getValueOfPatientPropertyFromResource({ resource });
+        // if we have any uuids then check if any of those are included in patient ids in patient scope
+        if (patientUuids && patientUuids.length > 0) {
+            if (patientUuids.includes(patientForResource)) {
+                return true;
+            }
+        }
+        // now check any non-uuids
+        const patientNonUuids = patientIds.filter(id => !isUuid(id));
+        if (patientNonUuids && patientNonUuids.length > 0) {
+            if (patientNonUuids.includes(patientForResource)) {
+                return true;
+            }
+        }
+        return false;
     }
 }
