+++ src/operations/common/securityTagManager.js
@@ -6,8 +6,6 @@ const { AccessIndexManager } = require('./accessIndexManager');
 const { SecurityTagSystem } = require('../../utils/securityTagSystem');
 const { PatientFilterManager } = require('../../fhir/patientFilterManager');
 const { isUuid } = require('../../utils/uid.util');
-const Resource = require('../../fhir/classes/4_0_0/resources/resource');
-const { NestedPropertyReader } = require('../../utils/nestedPropertyReader');
 
 /**
  * This class manages queries for security tags
@@ -32,6 +30,9 @@ class SecurityTagManager {
         this.accessIndexManager = accessIndexManager;
         assertTypeEquals(accessIndexManager, AccessIndexManager);
 
+        /**
+         * @type {PatientFilterManager}
+         */
         this.patientFilterManager = patientFilterManager;
         assertTypeEquals(patientFilterManager, PatientFilterManager);
     }
@@ -247,89 +248,6 @@ class SecurityTagManager {
             query = this.appendAndQuery(query, patientsQuery);
         }
         return query;
-    }
-
-    /**
-     * Gets value of patient property from resource
-     * @param {Resource} resource
-     * @return {string | undefined}
-     */
-    getValueOfPatientPropertyFromResource ({ resource }) {
-        assertTypeEquals(resource, Resource);
-        /** @type {string} */
-        const resourceType = resource.resourceType;
-        // Get name of patient property for this resource type
-        /**
-         * @type {string|string[]|null}
-         */
-        const patientFilterProperty = this.patientFilterManager.getPatientPropertyForResource({
-            resourceType
-        });
-        // If this resource has a patient property then get the value of that property
-        if (patientFilterProperty) {
-            if (Array.isArray(patientFilterProperty)) {
-                for (const p of patientFilterProperty) {
-                    // if patient itself then search by _uuid
-                    if (p === 'id') {
-                        return resource._uuid;
-                    } else {
-                        const propertyUuid = p.replace('.reference', '._uuid');
-                        const value = NestedPropertyReader.getNestedProperty({ obj: resource, path: propertyUuid });
-                        if (value !== undefined) {
-                            return value.replace('Patient/', '');
-                        }
-                    }
-                }
-            } else {
-                if (patientFilterProperty === 'id') {
-                    return resource._uuid;
-                } else {
-                    const propertyUuid = patientFilterProperty.replace('.reference', '._uuid');
-                    const value = NestedPropertyReader.getNestedProperty({ obj: resource, path: propertyUuid });
-                    if (value !== undefined) {
-                        return value.replace('Patient/', '');
-                    }
-                }
-            }
-        }
-        return undefined;
-    }
-
-    /**
-     * Gets whether the user can write the given resource with the given patient scope
-     * @param {string[] | null} patientIds
-     * @param {Resource} resource
-     * @return {boolean}
-     */
-    canWriteResourceWithPatientScope ({ patientIds, resource }) {
-        assertTypeEquals(resource, Resource);
-
-        /** @type {string} */
-        const resourceType = resource.resourceType;
-
-        // if this resource is not allowed to be written with a patient scope then return an error
-        if (!this.patientFilterManager.canAccessResourceWithPatientScope({ resourceType })) {
-            throw new ForbiddenError(`Resource type ${resourceType} cannot be written via a patient scope`);
-        }
-        // separate uuids from non-uuids
-        const patientUuids = patientIds.filter(id => isUuid(id));
-
-        /** @type {string} */
-        const patientForResource = this.getValueOfPatientPropertyFromResource({ resource });
-        // if we have any uuids then check if any of those are included in patient ids in patient scope
-        if (patientUuids && patientUuids.length > 0) {
-            if (patientUuids.includes(patientForResource)) {
-                return true;
-            }
-        }
-        // now check any non-uuids
-        const patientNonUuids = patientIds.filter(id => !isUuid(id));
-        if (patientNonUuids && patientNonUuids.length > 0) {
-            if (patientNonUuids.includes(patientForResource)) {
-                return true;
-            }
-        }
-        return false;
     }
 }
