+++ src/tests/operations/common/patientScopeManager/fixtures/Condition/condition1.json
