+++ src/tests/operations/common/patientScopeManager/fixtures/expected/expected_condition.json
