+++ src/tests/operations/common/patientScopeManager/patientScopeManager.test.js
@@ -13,7 +13,7 @@ const Patient = require('../../../../fhir/classes/4_0_0/resources/patient');
 const Condition = require('../../../../fhir/classes/4_0_0/resources/condition');
 const { generateUUIDv5 } = require('../../../../utils/uid.util');
 
-describe('SecurityTagManager Tests', () => {
+describe('PatientScopeManager Tests', () => {
     beforeEach(async () => {
         await commonBeforeEach();
     });
@@ -22,34 +22,34 @@ describe('SecurityTagManager Tests', () => {
         await commonAfterEach();
     });
 
-    describe('securityTagManager getValueOfPatientPropertyFromResource Tests', () => {
+    describe('patientScopeManager getValueOfPatientPropertyFromResource Tests', () => {
         test('getValueOfPatientPropertyFromResource works for Patient', async () => {
             /** @type {SimpleContainer} */
             const container = createTestContainer();
-            /** @type {SecurityTagManager} */
-            const securityTagManager = container.securityTagManager;
+            /** @type {PatientScopeManager} */
+            const patientScopeManager = container.patientScopeManager;
             /** @type {PreSaveManager} */
             const preSaveManager = container.preSaveManager;
             /** @type {Patient} */
             const patient = new Patient(patient1Resource);
             // generate all the uuids
             await preSaveManager.preSaveAsync(patient);
-            const patientUuid = securityTagManager.getValueOfPatientPropertyFromResource({ resource: patient });
+            const patientUuid = patientScopeManager.getValueOfPatientPropertyFromResource({ resource: patient });
             const expectedPatientUuid = generateUUIDv5(`${patient1Resource.id}|${patient1Resource.meta.security[0].code}`);
             expect(patientUuid).toStrictEqual(expectedPatientUuid);
         });
         test('getValueOfPatientPropertyFromResource works for Condition', async () => {
             /** @type {SimpleContainer} */
             const container = createTestContainer();
-            /** @type {SecurityTagManager} */
-            const securityTagManager = container.securityTagManager;
+            /** @type {PatientScopeManager} */
+            const patientScopeManager = container.patientScopeManager;
             /** @type {PreSaveManager} */
             const preSaveManager = container.preSaveManager;
             /** @type {Condition} */
             const condition = new Condition(condition1Resource);
             // generate all the uuids
             await preSaveManager.preSaveAsync(condition);
-            const patientUuid = securityTagManager.getValueOfPatientPropertyFromResource({ resource: condition });
+            const patientUuid = patientScopeManager.getValueOfPatientPropertyFromResource({ resource: condition });
             const patientReference = condition.subject.reference;
             const patientId = patientReference.split('/')[1];
             const expectedPatientUuid = generateUUIDv5(`${patientId}|${condition.meta.security[0].code}`);
@@ -57,12 +57,12 @@ describe('SecurityTagManager Tests', () => {
         });
     });
 
-    describe('securityTagManager canWriteResourceWithPatientScope Tests for Patient', () => {
+    describe('patientScopeManager canWriteResourceWithPatientScope Tests for Patient', () => {
         test('canWriteResourceWithPatientScope works for Patient', async () => {
             /** @type {SimpleContainer} */
             const container = createTestContainer();
-            /** @type {SecurityTagManager} */
-            const securityTagManager = container.securityTagManager;
+            /** @type {PatientScopeManager} */
+            const patientScopeManager = container.patientScopeManager;
             /** @type {PreSaveManager} */
             const preSaveManager = container.preSaveManager;
             /** @type {Patient} */
@@ -72,7 +72,7 @@ describe('SecurityTagManager Tests', () => {
             const patientUuid = generateUUIDv5(`${patient1Resource.id}|${patient1Resource.meta.security[0].code}`);
             // now do the test
             /** @type {boolean} */
-            const writeAllowed = securityTagManager.canWriteResourceWithPatientScope({
+            const writeAllowed = patientScopeManager.canWriteResourceWithPatientScope({
                 patientIds: [patientUuid],
                 resource: patient
             });
@@ -81,8 +81,8 @@ describe('SecurityTagManager Tests', () => {
         test('canWriteResourceWithPatientScope works with multiple uuids for Patient', async () => {
             /** @type {SimpleContainer} */
             const container = createTestContainer();
-            /** @type {SecurityTagManager} */
-            const securityTagManager = container.securityTagManager;
+            /** @type {PatientScopeManager} */
+            const patientScopeManager = container.patientScopeManager;
             /** @type {PreSaveManager} */
             const preSaveManager = container.preSaveManager;
             /** @type {Patient} */
@@ -92,7 +92,7 @@ describe('SecurityTagManager Tests', () => {
             const patientUuid = generateUUIDv5(`${patient1Resource.id}|${patient1Resource.meta.security[0].code}`);
             // now do the test
             /** @type {boolean} */
-            const writeAllowed = securityTagManager.canWriteResourceWithPatientScope({
+            const writeAllowed = patientScopeManager.canWriteResourceWithPatientScope({
                 patientIds: ['3d6d9c23-e357-465a-b7c3-6d177bcc27c7', patientUuid],
                 resource: patient
             });
@@ -101,8 +101,8 @@ describe('SecurityTagManager Tests', () => {
         test('canWriteResourceWithPatientScope fails with wrong uuid for Patient', async () => {
             /** @type {SimpleContainer} */
             const container = createTestContainer();
-            /** @type {SecurityTagManager} */
-            const securityTagManager = container.securityTagManager;
+            /** @type {PatientScopeManager} */
+            const patientScopeManager = container.patientScopeManager;
             /** @type {PreSaveManager} */
             const preSaveManager = container.preSaveManager;
             /** @type {Patient} */
@@ -112,7 +112,7 @@ describe('SecurityTagManager Tests', () => {
             const patientUuid = generateUUIDv5(`123|${patient1Resource.meta.security[0].code}`);
             // now do the test
             /** @type {boolean} */
-            const writeAllowed = securityTagManager.canWriteResourceWithPatientScope({
+            const writeAllowed = patientScopeManager.canWriteResourceWithPatientScope({
                 patientIds: [patientUuid],
                 resource: patient
             });
@@ -121,8 +121,8 @@ describe('SecurityTagManager Tests', () => {
         test('canWriteResourceWithPatientScope fails with multiple wrong uuid for Patient', async () => {
             /** @type {SimpleContainer} */
             const container = createTestContainer();
-            /** @type {SecurityTagManager} */
-            const securityTagManager = container.securityTagManager;
+            /** @type {PatientScopeManager} */
+            const patientScopeManager = container.patientScopeManager;
             /** @type {PreSaveManager} */
             const preSaveManager = container.preSaveManager;
             /** @type {Patient} */
@@ -133,7 +133,7 @@ describe('SecurityTagManager Tests', () => {
             const patientUuid2 = generateUUIDv5(`456|${patient1Resource.meta.security[0].code}`);
             // now do the test
             /** @type {boolean} */
-            const writeAllowed = securityTagManager.canWriteResourceWithPatientScope({
+            const writeAllowed = patientScopeManager.canWriteResourceWithPatientScope({
                 patientIds: [patientUuid, patientUuid2],
                 resource: patient
             });
@@ -141,12 +141,12 @@ describe('SecurityTagManager Tests', () => {
         });
     });
 
-    describe('securityTagManager canWriteResourceWithPatientScope Tests for Condition', () => {
+    describe('patientScopeManager canWriteResourceWithPatientScope Tests for Condition', () => {
         test('canWriteResourceWithPatientScope works for Condition', async () => {
             /** @type {SimpleContainer} */
             const container = createTestContainer();
-            /** @type {SecurityTagManager} */
-            const securityTagManager = container.securityTagManager;
+            /** @type {PatientScopeManager} */
+            const patientScopeManager = container.patientScopeManager;
             /** @type {PreSaveManager} */
             const preSaveManager = container.preSaveManager;
             /** @type {Condition} */
@@ -158,7 +158,7 @@ describe('SecurityTagManager Tests', () => {
             const patientUuid = generateUUIDv5(`${patientId}|${condition.meta.security[0].code}`);
             // now do the test
             /** @type {boolean} */
-            const writeAllowed = securityTagManager.canWriteResourceWithPatientScope({
+            const writeAllowed = patientScopeManager.canWriteResourceWithPatientScope({
                 patientIds: [patientUuid],
                 resource: condition
             });
@@ -167,8 +167,8 @@ describe('SecurityTagManager Tests', () => {
         test('canWriteResourceWithPatientScope works with multiple uuids for Patient', async () => {
             /** @type {SimpleContainer} */
             const container = createTestContainer();
-            /** @type {SecurityTagManager} */
-            const securityTagManager = container.securityTagManager;
+            /** @type {PatientScopeManager} */
+            const patientScopeManager = container.patientScopeManager;
             /** @type {PreSaveManager} */
             const preSaveManager = container.preSaveManager;
             /** @type {Condition} */
@@ -180,7 +180,7 @@ describe('SecurityTagManager Tests', () => {
             const patientUuid = generateUUIDv5(`${patientId}|${condition.meta.security[0].code}`);
             // now do the test
             /** @type {boolean} */
-            const writeAllowed = securityTagManager.canWriteResourceWithPatientScope({
+            const writeAllowed = patientScopeManager.canWriteResourceWithPatientScope({
                 patientIds: ['3d6d9c23-e357-465a-b7c3-6d177bcc27c7', patientUuid],
                 resource: condition
             });
@@ -189,8 +189,8 @@ describe('SecurityTagManager Tests', () => {
         test('canWriteResourceWithPatientScope fails with wrong uuid for Patient', async () => {
             /** @type {SimpleContainer} */
             const container = createTestContainer();
-            /** @type {SecurityTagManager} */
-            const securityTagManager = container.securityTagManager;
+            /** @type {PatientScopeManager} */
+            const patientScopeManager = container.patientScopeManager;
             /** @type {PreSaveManager} */
             const preSaveManager = container.preSaveManager;
             /** @type {Condition} */
@@ -200,7 +200,7 @@ describe('SecurityTagManager Tests', () => {
             const patientUuid = generateUUIDv5(`123|${condition.meta.security[0].code}`);
             // now do the test
             /** @type {boolean} */
-            const writeAllowed = securityTagManager.canWriteResourceWithPatientScope({
+            const writeAllowed = patientScopeManager.canWriteResourceWithPatientScope({
                 patientIds: [patientUuid],
                 resource: condition
             });
@@ -209,8 +209,8 @@ describe('SecurityTagManager Tests', () => {
         test('canWriteResourceWithPatientScope fails with multiple wrong uuid for Patient', async () => {
             /** @type {SimpleContainer} */
             const container = createTestContainer();
-            /** @type {SecurityTagManager} */
-            const securityTagManager = container.securityTagManager;
+            /** @type {PatientScopeManager} */
+            const patientScopeManager = container.patientScopeManager;
             /** @type {PreSaveManager} */
             const preSaveManager = container.preSaveManager;
             /** @type {Condition} */
@@ -221,7 +221,7 @@ describe('SecurityTagManager Tests', () => {
             const patientUuid2 = generateUUIDv5(`456|${patient1Resource.meta.security[0].code}`);
             // now do the test
             /** @type {boolean} */
-            const writeAllowed = securityTagManager.canWriteResourceWithPatientScope({
+            const writeAllowed = patientScopeManager.canWriteResourceWithPatientScope({
                 patientIds: [patientUuid, patientUuid2],
                 resource: condition
             });
